+++ wls-basisdaten-service/src/main/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/rest/handbuch/HandbuchController.java
@@ -2,8 +2,8 @@ package de.muenchen.oss.wahllokalsystem.basisdatenservice.rest.handbuch;
 
 import de.muenchen.oss.wahllokalsystem.basisdatenservice.exception.ExceptionConstants;
 import de.muenchen.oss.wahllokalsystem.basisdatenservice.rest.common.FileMapper;
-import de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.common.FileResponseEntityModel;
 import de.muenchen.oss.wahllokalsystem.basisdatenservice.rest.common.WahlbezirkArtDTO;
+import de.muenchen.oss.wahllokalsystem.basisdatenservice.service.common.FileResponseEntityModel;
 import de.muenchen.oss.wahllokalsystem.basisdatenservice.service.handbuch.HandbuchService;
 import de.muenchen.oss.wahllokalsystem.wls.common.exception.util.ExceptionFactory;
 import io.swagger.v3.oas.annotations.Operation;
