+++ wls-basisdaten-service/src/main/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/service/common/FileResponseEntityModel.java
@@ -1,4 +1,4 @@
-package de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.common;
+package de.muenchen.oss.wahllokalsystem.basisdatenservice.service.common;
 
 public record FileResponseEntityModel(byte[] responseBody, String headerContentType, String attachmentFilename) {
 }
